refactor: fix AuthService
